Feature/questions

+++ src/components/base/CardQuestion.tsx
@@ -5,6 +5,9 @@ import {
   Tooltip,
   Chip,
   Checkbox,
+  RadioGroup,
+  FormControlLabel,
+  TextField,
 } from "@mui/material";
 import Icon from "@mdi/react";
 import {
@@ -12,30 +15,44 @@ import {
   mdiCheckboxOutline,
   mdiLeadPencil,
   mdiTrashCan,
+  mdiPlus,
 } from "@mdi/js";
 
-import { showAlert } from "@/common";
+import { LEVEL_QUESTION, showAlert } from "@/common";
 import { IAnswer, QuestionType } from "@/features";
 import { OutputTiptap } from "./math/OutputTiptap";
+import { useLocation } from "react-router-dom";
+import { InputText } from "./InputText";
 interface Props {
   text: string;
-  answers: IAnswer[];
+  answers?: IAnswer[];
+  answerTF?: string;
+  answerShort?: string;
   type: QuestionType;
   index?: number;
   id: string;
-  handleDelete: (id: string) => void;
-  handleUpdate: () => void;
+  level?: string;
+  handleDelete?: (id: string) => void;
+  handleUpdate?: () => void;
+  handleAdd?: () => void;
+  typeButton?: string;
 }
 
 export const CardQuestion = ({
   type,
   text,
   answers,
+  answerTF,
+  answerShort,
   index,
   id,
+  level,
+  typeButton,
   handleDelete,
   handleUpdate,
+  handleAdd,
 }: Props) => {
+  const location = useLocation();
   const letters = ["A", "B", "C", "D"];
 
   const handleClickDelete = (id: string) => {
@@ -43,9 +60,24 @@ export const CardQuestion = ({
       title: "Bạn có chắc muốn xóa câu hỏi này",
     }).then((result) => {
       if (result.isConfirmed) {
-        return handleDelete(id);
+        if (handleDelete) {
+          return handleDelete(id);
+        }
       }
     });
+  };
+
+  const getTypeQuestion = () => {
+    switch (type) {
+      case QuestionType.SINGLE_CHOICE:
+        return "Trắc nghiệm";
+      case QuestionType.MULTIPLE_CHOICE:
+        return "Trắc nghiệm nhiều đáp án";
+      case QuestionType.TRUE_FALSE:
+        return "Trắc nghiệm Đúng/Sai";
+      case QuestionType.SHORT_ANSWER:
+        return "Câu trả lời ngắn";
+    }
   };
 
   return (
@@ -54,51 +86,69 @@ export const CardQuestion = ({
         <div className="mb-3 flex justify-between items-center">
           <div>
             <Chip
-              icon={
-                <Icon
-                  path={
-                    type === QuestionType.SINGLE_CHOICE
-                      ? mdiCheckCircleOutline
-                      : mdiCheckboxOutline
-                  }
-                  size={1}
-                />
-              }
-              label={
-                type === QuestionType.SINGLE_CHOICE
-                  ? "Single choice"
-                  : "Multiple choice"
-              }
+              label={getTypeQuestion()}
               variant="outlined"
               sx={{ color: "#1D8FE4", borderColor: "#1D8FE4", marginLeft: -1 }}
             />
             <Chip
-              label="1 điểm"
+              label={
+                LEVEL_QUESTION.find((i) => i?.id === level?.toString())?.label
+              }
               variant="outlined"
               sx={{ color: "#ed3a3a", borderColor: "#ed3a3a", marginLeft: 2 }}
             />
           </div>
           <div>
-            <Tooltip title="Sửa">
-              <IconButton
-                sx={{ color: "#e28d0f" }}
-                onClick={() => {
-                  handleUpdate();
-                }}
-              >
-                <Icon path={mdiLeadPencil} size={1} />
-              </IconButton>
-            </Tooltip>
-            <Tooltip title="Xóa">
-              <IconButton
-                sx={{ color: "#ED3A3A" }}
-                onClick={() => {
-                  handleClickDelete(id);
-                }}
-              >
-                <Icon path={mdiTrashCan} size={1} />
-              </IconButton>
-            </Tooltip>
+            {typeButton === "add" && (
+              <Tooltip title="Thêm">
+                <IconButton
+                  sx={{ color: "#e28d0f" }}
+                  onClick={() => {
+                    if (handleAdd) {
+                      handleAdd();
+                    }
+                  }}
+                >
+                  <Icon path={mdiPlus} size={1} />
+                </IconButton>
+              </Tooltip>
+            )}
+            {typeButton === "delete" && (
+              <Tooltip title="Xóa">
+                <IconButton
+                  sx={{ color: "#e28d0f" }}
+                  onClick={() => handleClickDelete(id)}
+                >
+                  <Icon path={mdiTrashCan} size={1} />
+                </IconButton>
+              </Tooltip>
+            )}
+            {!location.pathname.includes("/assignment") && (
+              <>
+                <Tooltip title="Sửa">
+                  <IconButton
+                    sx={{ color: "#e28d0f" }}
+                    onClick={() => {
+                      if (handleUpdate) {
+                        handleUpdate();
+                      }
+                    }}
+                  >
+                    <Icon path={mdiLeadPencil} size={1} />
+                  </IconButton>
+                </Tooltip>
+                <Tooltip title="Xóa">
+                  <IconButton
+                    sx={{ color: "#ED3A3A" }}
+                    onClick={() => {
+                      handleClickDelete(id);
+                    }}
+                  >
+                    <Icon path={mdiTrashCan} size={1} />
+                  </IconButton>
+                </Tooltip>
+              </>
+            )}
           </div>
         </div>
         <div className="text-base pb-4 flex">
@@ -108,16 +158,51 @@ export const CardQuestion = ({
           <OutputTiptap value={text} />
         </div>
         <div className="grid grid-cols-12 gap-4">
-          {answers.map((answer, index) => (
-            <div key={index} className="mt-6 col-span-3 flex items-center">
-              {type === QuestionType.SINGLE_CHOICE ? (
-                <Radio checked={answer.isCorrect} />
-              ) : (
-                <Checkbox checked={answer.isCorrect} />
-              )}
-              {letters[index]}. <OutputTiptap value={answer.text} />
-            </div>
-          ))}
+          {(type === QuestionType.SINGLE_CHOICE ||
+            type === QuestionType.MULTIPLE_CHOICE) &&
+            answers?.map((answer, index) => (
+              <div key={index} className="mt-6 col-span-3 flex items-center">
+                {type === QuestionType.SINGLE_CHOICE ? (
+                  <Radio checked={answer.isCorrect} />
+                ) : (
+                  // <Checkbox checked={answer.isCorrect} />
+                  <Checkbox checked={answer.isCorrect} />
+                )}
+                {letters[index]}. <OutputTiptap value={answer.text} />
+              </div>
+            ))}
+
+          {type === QuestionType.TRUE_FALSE && (
+            <RadioGroup row defaultValue={answerTF} className="mb-[20px]">
+              <FormControlLabel
+                sx={{
+                  "& .MuiTypography-root": {
+                    fontFamily: "'Be Vietnam Pro', sans-serif",
+                    marginRight: "20px",
+                  },
+                }}
+                value="Đúng"
+                control={<Radio />}
+                label="Đúng"
+                checked={answerTF === "Đúng"}
+              />
+              <FormControlLabel
+                sx={{
+                  "& .MuiTypography-root": {
+                    fontFamily: "'Be Vietnam Pro', sans-serif",
+                  },
+                }}
+                value="Sai"
+                control={<Radio />}
+                label="Sai"
+                checked={answerTF === "Sai"}
+              />
+            </RadioGroup>
+          )}
+
+          {type === QuestionType.SHORT_ANSWER && (
+            <TextField value={answerShort} />
+          )}
         </div>
       </div>
     </Card>
